sentiment_sum/requirements.txt
django>=4.0.0
transformers>=4.0.0
torch>=1.0.0
vaderSentiment>=3.3.2
scikit-learn>=1.0.0
numpy>=1.20.0
openai>=1.0.0
python-dotenv>=1.0.0
chardet>=5.0.0
pandas>=1.5.0 